dashboard_app/pyproject.toml
[tool.poetry]
name = "BachelorProject_Babiar_Dashboard_App"
version = "1.0.0"
description = ""
authors = ["Mário Babiar <mariobabiar11@gmail.com>"]
packages = [{ include = "dashboard_app" }]

[tool.poetry.dependencies]
dash = "^3.0.4"
dash-bootstrap-components = "^2.0.1"
flask-caching = "^2.3.1"
neo4j-rust-ext = "^5.28.1.0"
networkx = "^3.4.2"
pandas = "^2.2.3"
python = "^3.12"
pyvis = "^0.3.2"
termcolor = "^3.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
